# Make subsample faster in CPU