--- a/test/test_handlers.py
+++ b/test/test_handlers.py
@@ -623,12 +623,11 @@ def test_subsample_fn():
     def subsample_fn(rng_key):
         return numpyro.primitives._subsample_fn(size, subsample_size, rng_key)
 
-    # test that keys are not duplicated
-    for i in range(10):
-        x = subsample_fn(random.PRNGKey(i))
-        assert len(set(x)) == subsample_size
-
     rng_keys = random.split(random.PRNGKey(0), num_samples)
     subsamples = vmap(subsample_fn)(rng_keys)
-    i = random.randint(random.PRNGKey(1), (), 0, size)
-    assert_allclose(jnp.mean(subsamples == i, axis=0), jnp.full(subsample_size, 1 / size), atol=1e-3)
+    for k in range(1, 11):
+        i = random.randint(random.PRNGKey(k), (), 0, size)
+        assert_allclose(jnp.mean(subsamples == i, axis=0), jnp.full(subsample_size, 1 / size), atol=1e-3)
+
+        # test that values are not duplicated
+        assert len(set(subsamples[k])) == subsample_size