@fritzo 2021-01-07
Do you need to ensure double precision so that `random.uniform` will sample sets larger than 24e6?