@fehiepsi 2021-01-07
Thanks for reviewing and pointing that issue out, @fritzo! I have switched to `randint` instead of using `uniform`. It is a bit slower `451 µs ± 2.06 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)` but is still pretty fast and does not suffer from precision issue.